# openapi.yaml
openapi: 3.0.0
info:
  title: Cortex AI Memory Platform API
  version: 1.0.0
  description: API for managing short-term and long-term memories in the Cortex AI Memory Platform

paths:
  /api/memories:
    get:
      summary: Retrieve memories
      parameters:
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum: [short-term, long-term]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Memory'
    post:
      summary: Create a new memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [short-term, long-term]
                content:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'

  /api/memories/{id}:
    get:
      summary: Retrieve a specific memory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum: [short-term, long-term]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
    put:
      summary: Update a specific memory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [short-term, long-term]
                content:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
    delete:
      summary: Delete a specific memory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum: [short-term, long-term]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

components:
  schemas:
    Memory:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time